requirements.txt
infi.devicemanager
apscheduler